+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Changed grid highlighting logic to highlight a grid if either its parent or any of its children are to be highlighted
+- Added configuration options to change zoom levels and font sizes
 
 ### Removed
 - Removed title_size option for font size configuration and replaced with calculation based on zoom level